Moved bytes to a dedicated package
